[ADD] Refactor for login
