+++ client/app/components/login/login.scss
@@ -1,0 +1,7 @@
+.login{
+    background:red;
+    margin:auto;
+    padding:auto;
+    width:10000em;
+    border: 1px solid black;
+}
