--- client/app/styles/all.scss
@@ -1,1 +1,0 @@
-@import "login";
