--- client/app/styles/login.scss
@@ -1,6 +1,0 @@
-.login{
-    background:grey;
-    margin:auto;
-    padding:auto;
-    width:50em;
-}
